requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.1
dataclasses-json==0.5.14
dnspython==2.4.1
exceptiongroup==1.1.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
langsmith==0.0.21
lit==16.0.6
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pydantic==1.10.12
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.19
sympy==1.12
tenacity==8.2.2
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.31.0
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wsproto==1.2.0
yarl==1.9.2
